backend/pyproject.toml
[tool.poetry]
name = "eventos2"
version = "0.1.0"
description = ""
authors = [
	"yurihs <yurisalvador@hotmail.com>"
]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.2"
django-environ = "^0.4.5"
djangorestframework = "^3.10"
psycopg2-binary = "^2.8"
drf-yasg = "^1.16"
packaging = "^19.2"
Pillow = "^6.2.0"
gunicorn = "^19.9.0"
huey = "^2.1.2"
redis = "^3.3.11"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
flake8 = "^3.7"
pre-commit = "^1.18"
toml = "^0.10.0"
flake8-bugbear = "^19.8"
isort = "^4.3"
bandit = "^1.6"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
